requirements.txt
transformers
torch
datasets
accelerate
openai
google-generativeai
anthropic
deepseek
python-dotenv